API reference

RheumaComposites — Module

RheumaComposites

A Julia package for composite scores in Rheumatology.

Types

RheumaComposites.AbstractComposite — Type

AbstractComposite

Abstract type that specifies the category of composites.

It is either ContinuousComposite, BooleanComposite, or ModifiedComposite.

RheumaComposites.ContinuousComposite — Type

ContinuousComposite <: AbstractComposite

Abstract type for composites whose scores are floating point numbers.

See also score, BooleanComposite.

RheumaComposites.BooleanComposite — Type

BooleanComposite <: AbstractComposite

Abstract type for composites only implementing a definition of remission.

See also isremission, ContinuousComposite.

RheumaComposites.ModifiedComposite — Type

ModifiedComposite <: AbstractComposite

Abstract type representing alterations to the behaviour of existing composites by, e.g., changing remission thresholds.

RheumaComposites.DAS28 — Type

DAS28 <: ContinuousComposite

Abstract type for DAS28 subtypes.

See also DAS28ESR, DAS28CRP.

RheumaComposites.DAS28ESR — Type

DAS28ESR(; t28, s28, pga, apr)

Store the component measures of the DAS28ESR.

Components

t28 28 tender joint count

s28 28 swollen joint count

pga patient's global assessment

apr active phase reactant, here ESR

Units

Currently, pga must be a length (typically millimeters or centimeters) and apr must be a rate (typically mm/hr). See also Unitful.@u_str.

External links

DAS28 calculator https://www.4s-dawn.com/DAS28/

See also score, isremission, DAS28.

RheumaComposites.DAS28CRP — Type

DAS28CRP(; t28, s28, pga, apr)

Store the component measures of the DAS28CRP.

Components

t28 28 tender joint count

s28 28 swollen joint count

pga patient's global assessment

apr active phase reactant, here CRP

Units

Currently, pga must be a length (typically millimeters or centimeters) and crp must be a concentration (typically mg/dL or mg/L). See also Unitful.@u_str.

External links

The DAS28 calculator

See also score, isremission, DAS28.

RheumaComposites.SDAI — Type

SDAI(; t28, s28, pga, ega, crp)

Store component measures of the Simplified Disease Activity Index, or SDAI.

Components

t28 28 tender joint count

s28 28 swollen joint count

pga patient's global assessment

ega evaluator's global assessment

crp c-reactive protein

Units

Currently, pga and ega must be a length (typically millimeters or centimeters) and crp must be a concentration (typically mg/dL or mg/L). See also Unitful.@u_str.

External links

SDAI calculator

See also score, isremission.

RheumaComposites.CDAI — Type

CDAI(; t28, s28, pga, ega)

Store component measures of the Clinical Disease Activity Index, or CDAI.

Components

t28 28 tender joint count

s28 28 swollen joint count

pga patient's global assessment

ega evaluator's global assessment

Units

Currently, pga and ega must be a length (typically millimeters or centimeters). See also Unitful.@u_str.

External links

CDAI calculator

See also score, isremission.

RheumaComposites.Faceted — Type

Faceted{T} <: ModifiedComposite

This type indicates a further grouping of the stored composite's components.

RheumaComposites.BooleanRemission — Type

BooleanRemission(; t28, s28, pga, crp)

Store the components of the original ACR/EULAR Boolean remission.

Components

t28 28 tender joint count

s28 28 swollen joint count

pga patient's global assessment

crp C-reactive protein

Units

Currently, pga must be a length (typically millimeters or centimeters) and crp must be a concentration (typically mg/dL or mg/L). See also Unitful.@u_str.

See also isremission.

RheumaComposites.Revised — Type

Revised{T} <: ModifiedComposite

This type indicates that the stored composite's threshold for remission has been revised.

See also BooleanRemission.

RheumaComposites.ThreeItem — Type

ThreeItem{T} <: ModifiedComposite

This type indicates that BooleanRemission should ignore pga for determining remission.

RheumaComposites.AbstractComponent — Type

AbstractComponent

Abstract type for single-component outcomes.

See also isremission, PGA, SJC28.

RheumaComposites.PGA — Type

PGA <: AbstractComponent

This type represents the patient global assessment measured on the visual analogue scale.

See also value, SJC28, AbstractComponent.

RheumaComposites.SJC28 — Type

SJC28 <: AbstractComponent

This type represents the 28 swollen joints count.

See also value, PGA, AbstractComponent.

Interface

RheumaComposites.t28 — Function

Return the 28 tender-joint count.

RheumaComposites.s28 — Function

Return the 28 swollen-joint count.

RheumaComposites.pga — Function

Return the patient global assessment.

RheumaComposites.apr — Function

Return the acute phase reactant.

RheumaComposites.ega — Function

Return the evaluator's global assessment.

Return the evaluator's global assessment.

RheumaComposites.crp — Function

Return the CRP value.

Functions

RheumaComposites.intercept — Function

intercept(x::ContinuousComposite)

Return the intercept.

This function can be useful if you want to implement custom decomposition or component reweighting.

See also score, decompose.

RheumaComposites.weight — Function

weight(x::T) where {T}

Weight a composite score's components according to its weighting scheme.

Example

julia> DAS28CRP(t28=2, s28=2, pga=54u"mm", apr=19u"mg/L") |> weight
(0.7919595949289333, 0.39597979746446665, 0.756, 1.0784636184794367)

RheumaComposites.score — Function

score(c::ContinuousComposite; digits=3)

Score a composite and optionally specify the rounding precision.

Examples

julia> DAS28ESR(t28=4, s28=2, pga=64u"mm", apr=44u"mm/hr") |> score
5.061

RheumaComposites.isremission — Function

isremission(x::AbstractComposite)

Check whether a composite fulfils remission criteria.

Examples

julia> DAS28ESR(t28=4, s28=5, pga=44u"mm", apr=23u"mm/hr") |> isremission
false
julia> BooleanRemission(t28=1, s28=0, pga=14u"mm", crp=0.4u"mg/dl") |>
       revised |>
       isremission
true

RheumaComposites.decompose — Function

decompose(x::ContinuousComposite; digits=3)

Return the proportion to which each component contributes to the composite's score.

Optionally specify the number of digits to round the results to.

See also score.

Examples

julia> SDAI(t28=4, s28=5, pga=16u"mm", ega=12u"mm", crp=3u"mg/L") |> decompose
(t28 = 0.331, s28 = 0.413, pga = 0.132, ega = 0.099, crp = 0.025)

decompose(x::Faceted{<:ContinuousComposite}; digits=3)

Return the proportion to which each facet contributes to the composite's score.

Examples

julia> c0 = DAS28ESR(t28=4, s28=5, pga=14u"mm", apr=12u"mm/hr");

julia> faceted(c0, (objective=[:s28, :apr], subjective=[:t28, :pga])) |> decompose
(objective = 0.474, subjective = 0.525)

RheumaComposites.categorise — Function

categorise(x::ContinuousComposite)

Convert x to a discrete value.

Examples

julia> DAS28ESR(t28=4, s28=5, pga=12u"mm", apr=44u"mm/hr") |> categorise
"Moderate"

categorise(::Type{SDAI}, v)

Convert v to a discrete value using SDAI thresholds.

The same functionality exists for other ContinuousComposites.

See also DAS28ESR, DAS28CRP.

Examples

julia> categorise(SDAI, 3.6)
"Low"

RheumaComposites.faceted — Function

faceted(c0::ContinuousComposite, facets::NamedTuple)

Specify a custom grouping along which the composite can be analysed.

See also decompose.

RheumaComposites.revised — Function

revised(c0::AbstractComposite)

Change the calculation of scores or thresholds to use the revised definition of a composite.

Currently, revised versions are implemented only for BooleanRemission.

See also isremission.

RheumaComposites.threeitem — Function

threeitem(c0::BooleanRemission)

Change the calculation of Boolean remission to exclude patient global assessment.

See also isremission, BooleanRemission.

RheumaComposites.value — Function

Return the value stored in x.

Index

RheumaComposites

RheumaComposites.AbstractComponent

RheumaComposites.AbstractComposite

RheumaComposites.BooleanComposite

RheumaComposites.BooleanRemission

RheumaComposites.CDAI

RheumaComposites.ContinuousComposite

RheumaComposites.DAS28

RheumaComposites.DAS28CRP

RheumaComposites.DAS28ESR

RheumaComposites.Faceted

RheumaComposites.ModifiedComposite

RheumaComposites.PGA

RheumaComposites.Revised

RheumaComposites.SDAI

RheumaComposites.SJC28

RheumaComposites.ThreeItem

RheumaComposites.apr

RheumaComposites.categorise

RheumaComposites.crp

RheumaComposites.decompose

RheumaComposites.ega

RheumaComposites.faceted

RheumaComposites.intercept

RheumaComposites.isremission

RheumaComposites.pga

RheumaComposites.revised

RheumaComposites.s28

RheumaComposites.score

RheumaComposites.t28

RheumaComposites.threeitem

RheumaComposites.value

RheumaComposites.weight