API reference

RheumaComposites — Module

RheumaComposites

A Julia package for composite scores in Rheumatology.

Types

RheumaComposites.AbstractComposite — Type

AbstractComposite

Abstract type that specifies the category of composites.

It is either ContinuousComposite, BooleanComposite, or ModifiedComposite.

RheumaComposites.ContinuousComposite — Type

ContinuousComposite <: AbstractComposite

Abstract type for composites whose scores are floating point numbers.

See also score, BooleanComposite.

RheumaComposites.BooleanComposite — Type

BooleanComposite <: AbstractComposite

Abstract type for composites only implementing a definition of remission.

See also isremission, ContinuousComposite.

RheumaComposites.ModifiedComposite — Type

ModifiedComposite <: AbstractComposite

Abstract type representing alterations to existing composites.

See also faceted, revised, threeitem.

RheumaComposites.DAS28 — Type

DAS28 <: ContinuousComposite

Abstract type for DAS28 subtypes.

See also DAS28ESR, DAS28CRP.

RheumaComposites.DAS28ESR — Type

DAS28ESR(; tjc, sjc, pga, apr)

Store the component measures of the DAS28ESR.

Components

tjc 28 tender joint count

sjc 28 swollen joint count

pga patient's global assessment

apr active phase reactant, here ESR

Units

Currently, pga must be a length (typically millimeters or centimeters) and apr must be a rate (typically mm/hr). See also Unitful.@u_str.

Categories

$<$ 3.2 = Remission

$\leq$ 3.2 = Low

$\leq$ 5.1 = Moderate

$>$ 5.1 = High

External links

DAS28 calculator

See also score, categorise, isremission.

RheumaComposites.DAS28CRP — Type

DAS28CRP(; tjc, sjc, pga, apr)

Store the component measures of the DAS28CRP.

Components

tjc 28 tender joint count

sjc 28 swollen joint count

pga patient's global assessment

apr active phase reactant, here CRP

Units

Currently, pga must be a length (typically millimeters or centimeters) and crp must be a concentration (typically mg/dL or mg/L). See also Unitful.@u_str.

Categories

$<$ 2.9 = Remission

$\leq$ 2.9 = Low

$\leq$ 4.6 = Moderate

$>$ 4.6 = High

External links

DAS28 calculator

See also score, categorise, isremission.

RheumaComposites.SDAI — Type

SDAI(; tjc, sjc, pga, ega, crp)

Store component measures of the Simplified Disease Activity Index, or SDAI.

Components

tjc 28 tender joint count

sjc 28 swollen joint count

pga patient's global assessment

ega evaluator's global assessment

crp c-reactive protein

Units

Currently, pga and ega must be a length (typically millimeters or centimeters) and crp must be a concentration (typically mg/dL or mg/L). See also Unitful.@u_str.

Categories

$\leq$ 11.0 = Remission

$\leq$ 11.0 = Low

$\leq$ 26.0 = Moderate

$>$ 26.0 = High

External links

SDAI calculator

See also score, categorise, isremission.

RheumaComposites.CDAI — Type

CDAI(; tjc, sjc, pga, ega)

Store component measures of the Clinical Disease Activity Index, or CDAI.

Components

tjc 28 tender joint count

sjc 28 swollen joint count

pga patient's global assessment

ega evaluator's global assessment

Units

Currently, pga and ega must be a length (typically millimeters or centimeters). See also Unitful.@u_str.

Categories

$<$ 10.0 = Remission

$\leq$ 10.0 = Low

$\leq$ 22.0 = Moderate

$>$ 22.0 = High

External links

CDAI calculator

See also score, categorise, isremission.

RheumaComposites.Faceted — Type

Faceted{T} <: ModifiedComposite

This type indicates a further grouping of the stored composite's components.

RheumaComposites.BooleanRemission — Type

BooleanRemission(; tjc, sjc, pga, crp)

Store the components of the original ACR/EULAR Boolean remission.

Components

tjc 28 tender joint count

sjc 28 swollen joint count

pga patient's global assessment

crp C-reactive protein

Units

Currently, pga must be a length (typically millimeters or centimeters) and crp must be a concentration (typically mg/dL or mg/L). See also Unitful.@u_str.

See also isremission.

RheumaComposites.Revised — Type

Revised{T} <: ModifiedComposite

This type indicates that the stored composite's threshold for remission has been revised.

See also BooleanRemission.

RheumaComposites.Subset — Type

Subset{N,T}

Redefine a composite as a subset of its components.

RheumaComposites.AbstractComponent — Type

AbstractComponent

Abstract type for single-component outcomes.

See also isremission, PGA, SJC.

RheumaComposites.PGA — Type

PGA <: AbstractComponent

This type represents the patient global assessment measured on the visual analogue scale.

See also value, SJC, AbstractComponent.

RheumaComposites.SJC — Type

SJC <: AbstractComponent

This type represents the 28 swollen joints count.

See also value, PGA, AbstractComponent.

Interface

RheumaComposites.tjc — Function

tjc(x::AbstractComposite)

Return the 28 tender-joint count.

RheumaComposites.sjc — Function

sjc(x::AbstractComposite)

Return the 28 swollen-joint count.

RheumaComposites.pga — Function

pga(x::AbstractComposite)

Return the patient global assessment.

RheumaComposites.apr — Function

apr(x::AbstractComposite)

Return the acute phase reactant.

RheumaComposites.ega — Function

Return the evaluator's global assessment.

RheumaComposites.crp — Function

Return the CRP value.

RheumaComposites.offset — Function

offset(x::Revised{<:BooleanComposite})

Return the offsets to remission thresholds.

RheumaComposites.components — Function

components(x::AbstractComposite)

Return the fieldnames of the type of x.

Alias for fieldnames(typeof(x)).

components(x::ModifiedComposite)

Return the components of the unmodified composite.

components(x::Subset{N,<:BooleanComposite})

Return the components kept in the Subset.

RheumaComposites.root — Function

root(x::ModifiedComposite)

Return the unmodified composite.

Functions

RheumaComposites.intercept — Function

intercept(x::ContinuousComposite)

Return the intercept.

This function can be useful if you want to implement custom decomposition or component reweighting.

See also score, decompose.

RheumaComposites.weight — Function

weight(x::T) where {T}

Weight a composite score's components according to its weighting scheme.

Example

julia> DAS28CRP(tjc=2, sjc=2, pga=54u"mm", apr=19u"mg/L") |> weight
(0.7919595949289333, 0.39597979746446665, 0.756, 1.0784636184794367)

RheumaComposites.score — Function

score(c::ContinuousComposite; digits=3)

Score a composite and optionally specify the rounding precision.

Examples

julia> DAS28ESR(tjc=4, sjc=2, pga=64u"mm", apr=44u"mm/hr") |> score
5.061

RheumaComposites.isremission — Function

isremission(x::AbstractComposite)

Check whether a composite fulfils remission criteria.

Examples

julia> DAS28ESR(tjc=4, sjc=5, pga=44u"mm", apr=23u"mm/hr") |> isremission
false
julia> BooleanRemission(tjc=1, sjc=0, pga=14u"mm", crp=0.4u"mg/dl") |>
       revised |>
       isremission
true

RheumaComposites.decompose — Function

decompose(x::ContinuousComposite; digits=3)

Return the proportion to which each component contributes to the composite's score.

Optionally specify the number of digits to round the results to.

See also score.

Examples

julia> SDAI(tjc=4, sjc=5, pga=16u"mm", ega=12u"mm", crp=3u"mg/L") |> decompose
(tjc = 0.331, sjc = 0.413, pga = 0.132, ega = 0.099, crp = 0.025)

decompose(x::Faceted{<:ContinuousComposite}; digits=3)

Return the proportion to which each facet contributes to the composite's score.

Examples

julia> root = DAS28ESR(tjc=4, sjc=5, pga=14u"mm", apr=12u"mm/hr");

julia> faceted(root, (objective=[:sjc, :apr], subjective=[:tjc, :pga])) |> decompose
(objective = 0.474, subjective = 0.525)

RheumaComposites.categorise — Function

categorise(x::ContinuousComposite)

Convert x to a discrete value.

Examples

julia> DAS28ESR(tjc=4, sjc=5, pga=12u"mm", apr=44u"mm/hr") |> categorise
"Moderate"

categorise(::Type{SDAI}, v)

Convert v to a discrete value using SDAI thresholds.

The same functionality exists for other ContinuousComposites.

See also DAS28ESR, DAS28CRP.

Examples

julia> categorise(SDAI, 3.6)
"Low"

RheumaComposites.faceted — Function

faceted(root::ContinuousComposite, facets::NamedTuple)

Specify a custom grouping along which the composite can be analysed.

See also decompose.

RheumaComposites.revised — Function

revised(root::BooleanRemission, offset::NamedTuple)
revised(root::BooleanRemission; offset::NamedTuple=(; pga=10u"mm"))

Modify the remission thresholds of a composite.

The values passed to offset will be added to the default thresholds of root.

See also isremission.

RheumaComposites.subset — Function

subset(root::AbstractComposite, keep::Vector{Symbol})

Redefine a composite as a subset of its components.

Functions like score or isremission act on the subset of components.

RheumaComposites.threeitem — Function

threeitem(root::BooleanRemission)

Change the calculation of Boolean remission to exclude patient global assessment.

Alias for subset(root::BooleanRemission, [:sjc, :tjc, :crp]).

See also subset, isremission, BooleanRemission.

RheumaComposites.value — Function

Return the value stored in x.

Index

RheumaComposites

RheumaComposites.AbstractComponent

RheumaComposites.AbstractComposite

RheumaComposites.BooleanComposite

RheumaComposites.BooleanRemission

RheumaComposites.CDAI

RheumaComposites.ContinuousComposite

RheumaComposites.DAS28

RheumaComposites.DAS28CRP

RheumaComposites.DAS28ESR

RheumaComposites.Faceted

RheumaComposites.ModifiedComposite

RheumaComposites.PGA

RheumaComposites.Revised

RheumaComposites.SDAI

RheumaComposites.SJC

RheumaComposites.Subset

RheumaComposites.apr

RheumaComposites.categorise

RheumaComposites.components

RheumaComposites.crp

RheumaComposites.decompose

RheumaComposites.ega

RheumaComposites.faceted

RheumaComposites.intercept

RheumaComposites.isremission

RheumaComposites.offset

RheumaComposites.pga

RheumaComposites.revised

RheumaComposites.root

RheumaComposites.score

RheumaComposites.sjc

RheumaComposites.subset

RheumaComposites.threeitem

RheumaComposites.tjc

RheumaComposites.value

RheumaComposites.weight