API reference

RheumaComposites — Module

RheumaComposites

A Julia package for composite scores in Rheumatology.

Types

RheumaComposites.AbstractComposite — Type

AbstractComposite

Abstract type that specifies the category of composites.

It is either ContinuousComposite, BooleanComposite, or ModifiedComposite.

RheumaComposites.ContinuousComposite — Type

ContinuousComposite <: AbstractComposite

Abstract type for composites whose scores are floating point numbers.

See also score, BooleanComposite.

RheumaComposites.BooleanComposite — Type

BooleanComposite <: AbstractComposite

Abstract type for composites only implementing a definition of remission.

See also isremission, ContinuousComposite.

RheumaComposites.ModifiedComposite — Type

ModifiedComposite <: AbstractComposite

Abstract type representing alterations to existing composites.

See also faceted, revised, threeitem.

RheumaComposites.DAS28 — Type

DAS28 <: ContinuousComposite

Abstract type for DAS28 subtypes.

See also DAS28ESR, DAS28CRP.

RheumaComposites.DAS28ESR — Type

DAS28ESR(; tjc, sjc, pga, apr[; units])

Store the component measures of the DAS28ESR.

Optionally specify the units for each component using Unitful.@u_str.

Components

tjc 28 tender joint count

sjc 28 swollen joint count

pga (cm) patient's global assessment

apr (mm/hr) active phase reactant, here ESR

Categories

$<$ 3.2 = Remission

$\leq$ 3.2 = Low

$\leq$ 5.1 = Moderate

$>$ 5.1 = High

External links

DAS28 calculator

See also score, categorise, isremission.

RheumaComposites.DAS28CRP — Type

DAS28CRP(; tjc, sjc, pga, apr[; units])

Store the component measures of the DAS28CRP.

Optionally specify the units for each component using Unitful.@u_str.

Components

tjc 28 tender joint count

sjc 28 swollen joint count

pga (cm) patient's global assessment

apr (mg/L) active phase reactant, here CRP

Categories

$<$ 2.9 = Remission

$\leq$ 2.9 = Low

$\leq$ 4.6 = Moderate

$>$ 4.6 = High

External links

DAS28 calculator

See also score, categorise, isremission.

RheumaComposites.SDAI — Type

SDAI(; tjc, sjc, pga, ega, crp[; units])

Store component measures of the Simplified Disease Activity Index, or SDAI.

Optionally specify the units for each component using Unitful.@u_str.

Components

tjc 28 tender joint count

sjc 28 swollen joint count

pga (cm) patient's global assessment

ega (cm) evaluator's global assessment

crp (mg/dL) c-reactive protein

Categories

$\leq$ 11.0 = Remission

$\leq$ 11.0 = Low

$\leq$ 26.0 = Moderate

$>$ 26.0 = High

External links

SDAI calculator

See also score, categorise, isremission.

RheumaComposites.CDAI — Type

CDAI(; tjc, sjc, pga, ega[; units])

Store component measures of the Clinical Disease Activity Index, or CDAI.

Optionally specify the units for each component using Unitful.@u_str.

Components

tjc 28 tender joint count

sjc 28 swollen joint count

pga (cm) patient's global assessment

ega (cm) evaluator's global assessment

Categories

$<$ 10.0 = Remission

$\leq$ 10.0 = Low

$\leq$ 22.0 = Moderate

$>$ 22.0 = High

External links

CDAI calculator

See also score, categorise, isremission.

RheumaComposites.BASDAI — Type

BASDAI(; tjc, sjc, pga, jpn[; units])

Store component measures of the Bath Ankylosing Spondylitis Disease Activity Index, or BASDAI.

Optionally specify the units for each component using Unitful.@u_str.

Components

q1 (cm) fatigue or tiredness

q2 (cm) AS neck, back, or hip pain

q3 (cm) pain or swelling in other joints

q4 (cm) discomfort from tender areas

q5 (cm) morning stiffness severity

q6 (cm) morning stiffness duration

Categories

The only cutoff I am aware of checks if a person's BASDAI is below 4. This is currently implemented in categorise and isremission, making them effectively redundant. Feel free to open an issue on the GitHub page of this package if you know about other cutoffs.

External links

BASDAI calculator

See also score.

RheumaComposites.DAPSA — Type

DAPSA(; tjc, sjc, pga, jpn[; units])

Store component measures of the index for Disease Activity in Psoriatic Arthritis, or DAPSA.

Optionally specify the units for each component using Unitful.@u_str.

Components

tjc 66 tender joint count

sjc 68 swollen joint count

pga (cm) patient's global assessment

jpn (cm) joint pain

Categories

$\leq$ 14.0 = Remission

$\leq$ 14.0 = Low

$\leq$ 28.0 = Moderate

$>$ 28.0 = High

External links

DAPSA calculator

See also score, categorise, isremission.

RheumaComposites.Faceted — Type

Faceted{T} <: ModifiedComposite

This type indicates a further grouping of the stored composite's components.

RheumaComposites.BooleanRemission — Type

BooleanRemission(; tjc, sjc, pga, crp[; units])

Store the components of the original ACR/EULAR Boolean remission.

Optionally specify the units for each component using Unitful.@u_str.

Components

tjc 28 tender joint count

sjc 28 swollen joint count

pga (cm) patient's global assessment

crp (mg/dL) C-reactive protein

See also isremission.

RheumaComposites.Revised — Type

Revised{T} <: ModifiedComposite

This type indicates that the stored composite's threshold for remission has been revised.

See also BooleanRemission.

RheumaComposites.Partial — Type

Partial{T}

Redefine a composite as a subset of its components.

Interface

Base.values — Function

values(x::AbstractComposite)

Return the values stored in x.

values(x::ModifiedComposite)

Return the values of the unmodified composite, i.e., x.root.values.

values(x::Partial{N,<:BooleanComposite})

Return the values kept by x.

RheumaComposites.root — Function

root(x::ModifiedComposite)

Return the unmodified composite.

RheumaComposites.offsets — Function

offsets(x::Revised{<:BooleanComposite})

Return the offsets to remission thresholds.

Functions

RheumaComposites.intercept — Function

intercept(x::ContinuousComposite)

Return the intercept.

This function can be useful if you want to implement custom decomposition or component reweighting.

See also score, decompose.

RheumaComposites.weight — Function

weight(x::T) where {T}

Weight a composite score's components according to its weighting scheme.

Example

julia> DAS28CRP(tjc=2, sjc=2, pga=54, apr=19) |> weight
(0.7919595949289333, 0.39597979746446665, 0.756, 1.0784636184794367)

RheumaComposites.score — Function

score(c::ContinuousComposite; digits=3)

Score a composite and optionally specify the rounding precision.

Examples

julia> DAS28ESR(tjc=4, sjc=2, pga=64, apr=44) |> score
5.061

RheumaComposites.isremission — Function

isremission(::Type{T}, s::Real) where {T<:ContinuousComposite}

Check whether a composite fulfils remission criteria.

Examples

julia> isremission(DAS28ESR, 3.9)
false

isremission(x::AbstractComposite)

Check whether a composite fulfils remission criteria.

Examples

julia> DAS28ESR(tjc=4, sjc=5, pga=44, apr=23) |> isremission
false
julia> BooleanRemission(tjc=1, sjc=0, pga=1.4, crp=0.4) |>
       revised |>
       isremission
true

RheumaComposites.decompose — Function

decompose(x::ContinuousComposite; digits=3)

Return the proportion to which each component contributes to the composite's score.

Optionally specify the number of digits to round the results to.

See also score.

Examples

julia> SDAI(tjc=4, sjc=5, pga=1.6, ega=1.2, crp=3) |> decompose
Dict{Symbol, Float64} with 5 entries:
  :tjc => 0.27
  :ega => 0.081
  :sjc => 0.338
  :pga => 0.108
  :crp => 0.203

decompose(x::Faceted{<:ContinuousComposite}; digits=3)

Return the proportion to which each facet contributes to the composite's score.

Examples

julia> root = DAS28ESR(tjc=4, sjc=5, pga=14, apr=12);

julia> faceted(root, (objective=[:sjc, :apr], subjective=[:tjc, :pga])) |> decompose
Dict{Symbol, Float64} with 2 entries:
  :subjective => 0.357
  :objective  => 0.642

RheumaComposites.categorise — Function

categorise(::Type{T}, s::Real) where {T<:ContinuousComposite}

Convert score s to a discrete value using T's thresholds.

Examples

julia> categorise(SDAI, 3.6)
"low"

categorise(x::ContinuousComposite)

Convert x to a discrete value.

Examples

julia> DAS28ESR(tjc=4, sjc=5, pga=12, apr=44) |> categorise
"moderate"

categorise(x::Faceted{<:ContinuousComposite})

Convert the root composite of x to a discrete value.

RheumaComposites.faceted — Function

faceted(root::ContinuousComposite, facets::NamedTuple)

Specify a custom grouping along which the composite can be analysed.

See also decompose.

RheumaComposites.revised — Function

revised(root::BooleanRemission, offset::NamedTuple)
revised(root::BooleanRemission; offset::NamedTuple=(; pga=10u"mm"))

Modify the remission thresholds of a composite.

The values passed to offset will be added to the default thresholds of root.

See also isremission.

RheumaComposites.partial — Function

partial(root::AbstractComposite, keep::Vector{Symbol})

Redefine a composite as a subset of its components.

Functions like score or isremission act on the subset of components.

RheumaComposites.threeitem — Function

threeitem(root::BooleanRemission)

Change the calculation of Boolean remission to exclude patient global assessment.

Alias for partial(root::BooleanRemission, [:sjc, :tjc, :crp]).

See also partial, isremission, BooleanRemission.

Index

RheumaComposites

RheumaComposites.AbstractComposite

RheumaComposites.BASDAI

RheumaComposites.BooleanComposite

RheumaComposites.BooleanRemission

RheumaComposites.CDAI

RheumaComposites.ContinuousComposite

RheumaComposites.DAPSA

RheumaComposites.DAS28

RheumaComposites.DAS28CRP

RheumaComposites.DAS28ESR

RheumaComposites.Faceted

RheumaComposites.ModifiedComposite

RheumaComposites.Partial

RheumaComposites.Revised

RheumaComposites.SDAI

Base.values

RheumaComposites.categorise

RheumaComposites.decompose

RheumaComposites.faceted

RheumaComposites.intercept

RheumaComposites.isremission

RheumaComposites.offsets

RheumaComposites.partial

RheumaComposites.revised

RheumaComposites.root

RheumaComposites.score

RheumaComposites.threeitem

RheumaComposites.weight